don't overwrite AndroidManifest.xml and configAndroid.json
Don't create the generated configured files (AndroidManifest.xml and
configAndroid.json) in the binary directory because androiddeployqt will
overwrite them with the default ones.

We now generate AndroidManifest.xml and configAndroid.json in the source
directory and make "android-package-source-directory" point to
"@CMAKE_CURRENT_SOURCE_DIR@", this will avoid copying a bunch of files
as well.